% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temp_functions.R
\name{C_to_F}
\alias{C_to_F}
\title{Convert Celsius to Fahrenheit}
\usage{
C_to_F(C)
}
\arguments{
\item{C}{Numeric vector in Celsius}
}
\value{
Numeric vector in Fahrenheit
}
\description{
Convert Celsius to Fahrenheit
}
\examples{
C_to_F(100)
}
