% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temp_functions.R
\name{C_to_K}
\alias{C_to_K}
\title{Convert Celsius to Kelvin}
\usage{
C_to_K(C)
}
\arguments{
\item{C}{Numeric vector in Celsius}
}
\value{
Numeric vector in Kelvin
}
\description{
Convert Celsius to Kelvin
}
\examples{
C_to_K(0)
}
