% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temp_functions.R
\name{F_to_C}
\alias{F_to_C}
\title{Convert Fahrenheit to Celsius}
\usage{
F_to_C(F)
}
\arguments{
\item{F}{Numeric vector in Fahrenheit}
}
\value{
Numeric vector in Celsius
}
\description{
Convert Fahrenheit to Celsius
}
\examples{
F_to_C(212)
}
