% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temp_functions.R
\name{F_to_K}
\alias{F_to_K}
\title{Convert Fahrenheit to Kelvin}
\usage{
F_to_K(F)
}
\arguments{
\item{F}{Numeric vector in Fahrenheit}
}
\value{
Numeric vector in Kelvin
}
\description{
Convert Fahrenheit to Kelvin
}
\examples{
F_to_K(32)
}
