% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temp_functions.R
\name{K_to_C}
\alias{K_to_C}
\title{Convert Kelvin to Celsius}
\usage{
K_to_C(K)
}
\arguments{
\item{K}{Numeric vector in Kelvin}
}
\value{
Numeric vector in Celsius
}
\description{
Convert Kelvin to Celsius
}
\examples{
K_to_C(273.15)
}
