% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temp_functions.R
\name{K_to_F}
\alias{K_to_F}
\title{Convert Kelvin to Fahrenheit}
\usage{
K_to_F(K)
}
\arguments{
\item{K}{Numeric vector in Kelvin}
}
\value{
Numeric vector in Fahrenheit
}
\description{
Convert Kelvin to Fahrenheit
}
\examples{
K_to_F(373.15)
}
